convert src files to typescript
