+++ .env-example
@@ -1,4 +1,4 @@
-NODE_ENV=dev
+NODE_ENV=production
 LOG_LEVEL=error
 LOG_FILE=/var/log/clive.log
 DB_FILE=/srv/clive/db.json
