+++ .eslintrc
@@ -12,8 +12,16 @@
       }
     ]
   },
+  "parser": "@babel/eslint-parser",
   "parserOptions": {
-    "ecmaVersion": 2019
+    "ecmaVersion": 2022,
+    "requireConfigFile": false,
+    "babelOptions": {
+      "babelrc": false,
+      "configFile": false,
+      // your babel options
+      // "presets": ["@babel/preset-env"]
+    }
   },
   "env": {
     "node": true,
