+++ package.json
@@ -1,32 +1,39 @@
 {
 	"name": "clappybot",
-	"version": "1.0.0",
+	"version": "1.1.0",
 	"description": "bot that posts clips posted to twitch chat to discord",
 	"license": "ISC",
-	"main": "index.js",
+	"exports": "./index.js",
 	"scripts": {
-		"start": "nodemon src/index.js",
-		"dev": "nodemon --inspect src/index.js",
+		"build": "tsc",
+		"start": "npm run build && node build/index.js",
+		"dev:build": "tsc -w",
+		"dev:run": "nodemon --inspect build/index.js",
+		"dev": "concurrently npm:dev:*",
 		"precommit": "lint-staged"
 	},
 	"dependencies": {
+		"axios": "^1.6.8",
 		"dotenv": "^16.4.5",
-		"lodash": "^4.17.21",
 		"lowdb": "^7.0.1",
-		"nodemon": "^3.1.0",
 		"request": "^2.88.2",
-		"request-promise": "^4.2.5",
 		"twitch-js": "^2.0.0-beta.45",
 		"winston": "^3.13.0"
 	},
 	"devDependencies": {
+		"@babel/core": "^7.24.4",
+		"@babel/eslint-parser": "^7.24.1",
+		"@types/node": "^20.12.4",
+		"concurrently": "^8.2.2",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-prettier": "^5.1.3",
 		"husky": "^9.0.11",
 		"lint-staged": "^15.2.2",
+		"nodemon": "^3.1.0",
 		"prettier": "^3.2.5",
-		"prettier-eslint": "^16.3.0"
+		"prettier-eslint": "^16.3.0",
+		"typescript": "^5.4.3"
 	},
 	"lint-staged": {
 		"**/*.js": [
@@ -38,5 +45,9 @@
 		"**/*.md": [
 			"prettier --write --parser --prose-wrap never markdown"
 		]
+	},
+	"type": "module",
+	"engines": {
+		"node": ">=20"
 	}
 }
